Former-commit-id: a669337266b0b101b5c3caa5b7a45e4d1e509857

+++ libraries/parallel/mpi_reconstruct_significant.cpp
@@ -25,6 +25,16 @@
 
 #include "mpi_reconstruct_significant.h"
 
+MpiProgReconstructSignificant::MpiProgReconstructSignificant()
+{
+	node=NULL;
+}
+
+MpiProgReconstructSignificant::~MpiProgReconstructSignificant()
+{
+	delete node;
+}
+
 void MpiProgReconstructSignificant::read(int argc, char** argv)
 {
     node = new MpiNode(argc, argv);
@@ -53,14 +63,17 @@ void MpiProgReconstructSignificant::gatherAlignment()
 		cc=aux;
 
 	// Write all metadatas
-	for (size_t n=0; n<mdReconstructionPartial.size(); ++n)
-	{
-		FileName fnPartial=formatString("%s/partial_node%03d_%03d.xmd",fnDir.c_str(),(int)rank,(int)n);
-		FileName fnProjectionMatching=formatString("%s/projmatch_node%03d_%03d.xmd",fnDir.c_str(),(int)rank,(int)n);
+	if (rank!=0)
+		for (size_t n=0; n<mdReconstructionPartial.size(); ++n)
+		{
+			FileName fnPartial=formatString("%s/partial_node%03d_%03d.xmd",fnDir.c_str(),(int)rank,(int)n);
+			FileName fnProjectionMatching=formatString("%s/projmatch_node%03d_%03d.xmd",fnDir.c_str(),(int)rank,(int)n);
 
-		mdReconstructionPartial[n].write(fnPartial);
-		mdReconstructionProjectionMatching[n].write(fnProjectionMatching);
-	}
+			if (mdReconstructionPartial[n].size()>0)
+				mdReconstructionPartial[n].write(fnPartial);
+			if (mdReconstructionProjectionMatching[n].size()>0)
+				mdReconstructionProjectionMatching[n].write(fnProjectionMatching);
+		}
 	synchronize();
 
 	// Now the master takes all of them
@@ -74,14 +87,23 @@ void MpiProgReconstructSignificant::gatherAlignment()
 				FileName fnPartial=formatString("%s/partial_node%03d_%03d.xmd",fnDir.c_str(),(int)otherRank,(int)n);
 				FileName fnProjectionMatching=formatString("%s/projmatch_node%03d_%03d.xmd",fnDir.c_str(),(int)otherRank,(int)n);
 
-				MDAux.read(fnPartial);
-				mdReconstructionPartial[n].unionAll(MDAux);
-				MDAux.read(fnProjectionMatching);
-				mdReconstructionProjectionMatching[n].unionAll(MDAux);
-				deleteFile(fnPartial);
-				deleteFile(fnProjectionMatching);
+				if (fnPartial.exists())
+				{
+					MDAux.read(fnPartial);
+					mdReconstructionPartial[n].unionAll(MDAux);
+					deleteFile(fnPartial);
+				}
+				if (fnProjectionMatching.exists())
+				{
+					MDAux.read(fnProjectionMatching);
+					mdReconstructionProjectionMatching[n].unionAll(MDAux);
+					deleteFile(fnProjectionMatching);
+				}
 			}
 		}
 	}
+
+	// std::cout << "synchronize" << std::endl;
 	synchronize();
+
 }
